glue example together
